requirements.txt
# Core ML dependencies
torch>=1.9.0
torchvision>=0.10.0
transformers
datasets>=2.0.0
tokenizers>=0.12.0

# Training utilities
accelerate>=0.20.0
scikit-learn>=1.0.0

# Experiment tracking
wandb>=0.15.0

# Configuration management
PyYAML>=6.0
argparse

# Data processing
pandas>=1.3.0
numpy>=1.21.0

# Image processing
Pillow>=8.0.0

# Development tools
tqdm>=4.62.0
jax
joblib
